+++ resources/views/pages/paciente/_forms_paciente/diagnostico_covid.blade.php
